create FILTER_BY_SIZE, which will be used in the very compact EH format.
The very compact EH format uses the return address as a key,
meaning lookups are literally constant time. This wastes some
space with padding, but also should dramatically speedup lookup
time. Combined with the very compact unwinding format, and this
should be orders of magnitude better than the status quo.